- OneChannelOrderPerOutputSMCAlgorithm  - tParticleCandidate struct defined inside the class (rather than inside a particular method)  - implemented restriction (old way commented) of a fixed number of survivors per state   - not thoroughly tested though - ResamplingAlgorithm  - obtainIndexes   - taken into account the case in which there are no trues in the mask   - fixed BUG - PSPBasedSMCAlgorithm  - added code to keep a number of survivors per state...though commented - StatUtil  - added some static_cast to remove compiler warnings

+++ statistics/StatUtil.cpp
@@ -261,7 +261,7 @@ inline double StatUtil::computeFromActiveOperands(const VectorXd &probabilities,
 
 double StatUtil::probApriori(const VectorXd &symbolsVector, const std::vector<UsersActivityDistribution> &symbolsDistributions)
 {
-  if(symbolsVector.size()!=symbolsDistributions.size())
+  if(static_cast<uint>(symbolsVector.size())!=symbolsDistributions.size())
 	throw RuntimeException("StatUtil::probApriori: the number of symbols in the vector and that of distributions don't match.");
   
   double res = 1.0;
@@ -278,7 +278,7 @@ double StatUtil::probXgivenY(VectorXd &X, VectorXd &Y, const std::vector<UsersAc
   if(X.size()!=Y.size())
 	throw RuntimeException("StatUtil::probXgivenY: the sizes of the vectors don't match.");
 
-  if(X.size()!=symbolsDistributions.size())
+  if(static_cast<uint>(X.size())!=symbolsDistributions.size())
 	throw RuntimeException("StatUtil::probXgivenY: the number of symbols in the vectors and that of distributions don't match.");  
   
   double res = 1.0;
